rename ruckig smoothing files
