+++ moveit_core/trajectory_processing/include/moveit/trajectory_processing/ruckig_traj_smoothing.h
